+++ internal/server/http/server.go
@@ -1,4 +1,4 @@
-package server
+package http
 
 import (
 	"github.com/DucTran999/auth-service/config"
