+++ internal/server/http/validator.go
@@ -1,4 +1,4 @@
-package server
+package http
 
 import (
 	"regexp"
